requirements.txt
google-generativeai
moviepy==1.0.3
edge-tts
pillow
requests
numpy
opencv-python
imageio[pyav]
imageio[ffmpeg]
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
